assert: allow non-int returning functions to assert
Include GIT_ASSERT_EXT and GIT_ASSERT_ARG_EXT so that functions that do
not return int (or more precisely, where `-1` would not be an error code)
can assert.

This allows functions that return, eg, NULL on an error code to do that
by passing the return value (in this example, `NULL`) as a second
parameter to the GIT_ASSERT_EXT functions.